# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API Documentation",
        "description": "API documentation for Point of Sale application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/billing/subscribe": {
            "post": {
                "tags": [
                    "Billing"
                ],
                "summary": "Subscribe to PRO plan via Midtrans",
                "operationId": "e54f227aa47c4933a8e80699684b7346",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BillingSubscribeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "subscription": {
                                            "$ref": "#/components/schemas/Subscription"
                                        },
                                        "payment": {
                                            "$ref": "#/components/schemas/Payment"
                                        },
                                        "payment_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/billing/webhook": {
            "post": {
                "tags": [
                    "Billing"
                ],
                "summary": "Midtrans payment webhook (public)",
                "operationId": "11555829615560b0bb90fdb32e18b7f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "string"
                                    },
                                    "transaction_status": {
                                        "type": "string"
                                    },
                                    "payment_type": {
                                        "type": "string"
                                    },
                                    "transaction_time": {
                                        "type": "string"
                                    },
                                    "gross_amount": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed"
                    },
                    "400": {
                        "description": "Invalid notification"
                    }
                }
            }
        },
        "/billing/active": {
            "get": {
                "tags": [
                    "Billing"
                ],
                "summary": "Get active subscription",
                "operationId": "7a52f58c1dcd78a8f892f313b80b2a71",
                "responses": {
                    "200": {
                        "description": "Active subscription data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Subscription"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BillingSubscribeRequest": {
                "required": [
                    "plan_name"
                ],
                "properties": {
                    "plan_name": {
                        "type": "string",
                        "example": "PRO",
                        "enum": [
                            "PRO"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Subscription": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_name": {
                        "type": "string",
                        "example": "PRO"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29000
                    },
                    "duration_days": {
                        "type": "integer",
                        "example": 30
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "ACTIVE",
                            "EXPIRED"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subscription_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "string",
                        "example": "ordered-uuid"
                    },
                    "gross_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 29000
                    },
                    "payment_type": {
                        "type": "string",
                        "example": "MIDTRANS"
                    },
                    "transaction_time": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "SETTLEMENT",
                            "EXPIRED",
                            "CANCEL",
                            "DENY"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}